+++ core/config/config.py
@@ -8,6 +8,7 @@ from dataclasses_json import CatchAll, DataClassJsonMixin, Undefined, dataclass_
 from diffusers.schedulers.scheduling_utils import KarrasDiffusionSchedulers
 
 from core.config.samplers.sampler_config import SamplerConfig
+from core.flags import HighResFixFlag
 from core.types import SigmaScheduler
 
 logger = logging.getLogger(__name__)
@@ -188,6 +189,13 @@ class APIConfig:
     ] = "lllyasviel/Fooocus-Expansion"
     prompt_to_prompt_device: Literal["cpu", "gpu"] = "gpu"
 
+    # Free U
+    free_u: bool = False
+    free_u_s1: float = 0.9
+    free_u_s2: float = 0.2
+    free_u_b1: float = 1.2
+    free_u_b2: float = 1.4
+
     @property
     def dtype(self):
         "Return selected data type"
@@ -263,6 +271,13 @@ class FrontendConfig:
     disable_analytics: bool = False
 
 
+@dataclass
+class FlagsConfig:
+    "Configuration for flags"
+
+    highres: HighResFixFlag = field(default_factory=HighResFixFlag)
+
+
 @dataclass_json(undefined=Undefined.INCLUDE)
 @dataclass
 class Configuration(DataClassJsonMixin):
@@ -279,6 +294,7 @@ class Configuration(DataClassJsonMixin):
     onnx: ONNXConfig = field(default_factory=ONNXConfig)
     bot: BotConfig = field(default_factory=BotConfig)
     frontend: FrontendConfig = field(default_factory=FrontendConfig)
+    flags: FlagsConfig = field(default_factory=FlagsConfig)
     sampler_config: SamplerConfig = field(default_factory=SamplerConfig)
     extra: CatchAll = field(default_factory=dict)