+++ core/flags.py
@@ -23,8 +23,16 @@ class Flag:
 class HighResFixFlag(Flag, DataClassJsonMixin):
     "Flag to fix high resolution images"
 
-    scale: float
-    latent_scale_mode: LatentScaleModel = "bilinear"
+    scale: float = 2
+    mode: Literal["latent", "image"] = "latent"
+
+    # Image Upscaling
+    image_upscaler: str = "RealESRGAN_x4plus_anime_6B"
+
+    # Latent Upscaling
+    latent_scale_mode: LatentScaleModel = "bislerp-tortured"
+    antialiased: bool = False
+
+    # Img2img
     strength: float = 0.7
     steps: int = 50
-    antialiased: bool = False