+++ core/queue.py
@@ -1,5 +1,5 @@
-import asyncio
 import logging
+import threading
 from typing import List
 
 logger = logging.getLogger(__name__)
@@ -10,14 +10,14 @@ class Queue:
 
     def __init__(self) -> None:
         self.jobs: List[str] = []
-        self.lock = asyncio.Lock()
-        self.condition = asyncio.Condition(self.lock)
+        self.lock = threading.Lock()
+        self.condition = threading.Condition(self.lock)
         self.concurrent_jobs = 1  # Broken, do not change (torch becomes locked up)
 
-    async def mark_finished(self, job_id: str):
+    def mark_finished(self, job_id: str):
         "Mark the current job as finished"
 
-        async with self.lock:
+        with self.lock:
             try:
                 self.jobs.remove(job_id)
             except ValueError:
@@ -28,14 +28,14 @@ class Queue:
             self.condition.notify_all()
             logger.info(f"Job {job_id} has been processed")
 
-    async def wait_for_turn(self, job_id: str):
+    def wait_for_turn(self, job_id: str):
         "Wait until the job can be processed"
 
-        async with self.lock:
+        with self.lock:
             self.jobs.append(job_id)
 
             while job_id not in self.jobs[: self.concurrent_jobs]:
-                await self.condition.wait()
+                self.condition.wait()
 
             logger.info(f"Job {job_id} is now being processed")
             return
@@ -46,9 +46,9 @@ class Queue:
         self.jobs = []
         logger.info("Queue has been cleared")
 
-    async def trigger_condition(self):
+    def trigger_condition(self):
         "Trigger the condition manually, this will cause all jobs to check if they can be processed"
 
-        async with self.lock:
+        with self.lock:
             self.condition.notify_all()
             logger.info("Queue Condition has been triggered manually")